@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <RootNamespace>wan24.Crypto.Tests</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <AssemblyName>wan24CryptoSharedTests</AssemblyName>
     <PackageId>wan24-Crypto-Shared-Tests</PackageId>
     <Title>wan24-Crypto Shared Tests</Title>
-    <Version>1.13.0</Version>
+    <Version>2.0.0</Version>
     <Authors>nd1012</Authors>
     <Company>Andreas Zimmermann, wan24.de</Company>
     <Product>wan24-Crypto</Product>
