@@ -5,7 +5,11 @@ namespace wan24.Crypto
     /// <summary>
     /// Havege RNG
     /// </summary>
-    public sealed class HavegeRng : DisposableRngBase
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="capacity">Random data stream pool capacity</param>
+    public sealed class HavegeRng(in int capacity) : DisposableRngBase()
     {
         /// <summary>
         /// Havege RNG command
@@ -15,26 +19,19 @@ namespace wan24.Crypto
         /// <summary>
         /// Havege CLI arguments
         /// </summary>
-        private static readonly string[] HavegeArgs = new string[] { "-n", "0" };
+        private static readonly string[] HavegeArgs = ["-n", "0"];
 
         /// <summary>
         /// Havege stream pool
         /// </summary>
-        private readonly DisposableObjectPool<ProcessStream> Pool;
-
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="capacity">Random data stream pool capacity</param>
-        public HavegeRng(in int capacity) : base() => Pool = new(capacity, () => ProcessStream.Create(HAVEGE, args: HavegeArgs));
+        private readonly DisposableObjectPool<ProcessStream> Pool = new(capacity, () => ProcessStream.Create(HAVEGE, args: HavegeArgs));
 
         /// <inheritdoc/>
         public override Span<byte> FillBytes(in Span<byte> buffer)
         {
             EnsureUndisposed();
             using RentedObject<ProcessStream> havege = new(Pool);
-            int red = havege.Object.Read(buffer);
-            if (red != buffer.Length) throw new IOException($"Failed to read {red} byte from haveg process stream");
+            havege.Object.ReadExactly(buffer);
             return buffer;
         }
 
@@ -43,8 +40,7 @@ namespace wan24.Crypto
         {
             EnsureUndisposed();
             using RentedObject<ProcessStream> havege = new(Pool);
-            int red = await havege.Object.ReadAsync(buffer, cancellationToken).DynamicContext();
-            if (red != buffer.Length) throw new IOException($"Failed to read {red} byte from haveg process stream");
+            await havege.Object.ReadExactlyAsync(buffer, cancellationToken).DynamicContext();
             return buffer;
         }
 
