@@ -6,35 +6,28 @@
         /// <summary>
         /// PAKE event arguments
         /// </summary>
-        public sealed class PakeEventArgs : EventArgs
+        /// <remarks>
+        /// Constructor
+        /// </remarks>
+        /// <param name="context">Context</param>
+        /// <param name="pake">PAKE instance</param>
+        /// <param name="pakeServerEventArgs">Pake server event arguments of the original event</param>
+        public sealed class PakeEventArgs(ServerAuthContext context, Pake pake, Pake.PakeServerEventArgs pakeServerEventArgs) : EventArgs()
         {
-            /// <summary>
-            /// Constructor
-            /// </summary>
-            /// <param name="context">Context</param>
-            /// <param name="pake">PAKE instance</param>
-            /// <param name="pakeServerEventArgs">Pake server event arguments of the original event</param>
-            public PakeEventArgs(ServerAuthContext context, Pake pake, Pake.PakeServerEventArgs pakeServerEventArgs) : base()
-            {
-                Context = context;
-                Pake = pake;
-                PakeServerEventArgs = pakeServerEventArgs;
-            }
-
             /// <summary>
             /// Context
             /// </summary>
-            public ServerAuthContext Context { get; }
+            public ServerAuthContext Context { get; } = context;
 
             /// <summary>
             /// PAKE instance
             /// </summary>
-            public Pake Pake { get; }
+            public Pake Pake { get; } = pake;
 
             /// <summary>
             /// Pake server event arguments of the original event
             /// </summary>
-            public Pake.PakeServerEventArgs PakeServerEventArgs { get; }
+            public Pake.PakeServerEventArgs PakeServerEventArgs { get; } = pakeServerEventArgs;
         }
     }
 }
