@@ -5,12 +5,11 @@ namespace wan24.Crypto
     /// <summary>
     /// <c>/dev/random</c> readable stream pool
     /// </summary>
-    public sealed class DevRandomStreamPool : DisposableObjectPool<Stream>
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="capacity">Capacity</param>
+    public sealed class DevRandomStreamPool(in int capacity) : DisposableObjectPool<Stream>(capacity, RND.GetDevRandom)
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="capacity">Capacity</param>
-        public DevRandomStreamPool(in int capacity) : base(capacity, RND.GetDevRandom) { }
     }
 }
