@@ -3,27 +3,22 @@
     /// <summary>
     /// Event arguments for the <see cref="CryptoHelper.OnForcePostQuantum"/> event
     /// </summary>
-    public sealed class ForcePostQuantumEventArgs : EventArgs
+    /// <remarks>
+    /// Constructor
+    /// </remarks>
+    /// <param name="firstCall">Is the first call?</param>
+    /// <param name="strict">Strict post quantum-safety?</param>
+    public sealed class ForcePostQuantumEventArgs(bool firstCall, bool strict) : EventArgs()
     {
-        /// <summary>
-        /// Constructor
-        /// </summary>
-        /// <param name="firstCall">Is the first call?</param>
-        /// <param name="strict">Strict post quantum-safety?</param>
-        public ForcePostQuantumEventArgs(bool firstCall, bool strict) : base()
-        {
-            FirstCall = firstCall;
-            Strict = strict;
-        }
 
         /// <summary>
         /// Is the first call?
         /// </summary>
-        public bool FirstCall { get; }
+        public bool FirstCall { get; } = firstCall;
 
         /// <summary>
         /// Strict post quantum-safety?
         /// </summary>
-        public bool Strict { get; }
+        public bool Strict { get; } = strict;
     }
 }
