@@ -22,7 +22,7 @@ namespace wan24.Crypto
              * NOTE: This piece of code is almost a 1:1 copy of the RandomNumberGenerator code, which is licensed under the MIT license by the .NET Foundation. See 
              * Rng.LICENSE.md for details.
              */
-            if (fromInclusive >= toExclusive) throw new ArgumentOutOfRangeException(nameof(toExclusive));
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(fromInclusive, toExclusive);
             uint range = (uint)toExclusive - (uint)fromInclusive - 1;
             if (range == 0) return fromInclusive;
             uint mask = range;
@@ -70,7 +70,7 @@ namespace wan24.Crypto
              * NOTE: This piece of code is almost a 1:1 copy of the RandomNumberGenerator code, which is licensed under the MIT license by the .NET Foundation. See 
              * Rng.LICENSE.md for details.
              */
-            if (fromInclusive >= toExclusive) throw new ArgumentOutOfRangeException(nameof(toExclusive));
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(fromInclusive, toExclusive);
             uint range = (uint)toExclusive - (uint)fromInclusive - 1;
             if (range == 0) return fromInclusive;
             uint mask = range;
@@ -119,7 +119,7 @@ namespace wan24.Crypto
              * NOTE: This piece of code is almost a 1:1 copy of the RandomNumberGenerator code, which is licensed under the MIT license by the .NET Foundation. See 
              * Rng.LICENSE.md for details.
              */
-            if (fromInclusive >= toExclusive) throw new ArgumentOutOfRangeException(nameof(toExclusive));
+            ArgumentOutOfRangeException.ThrowIfGreaterThanOrEqual(fromInclusive, toExclusive);
             uint range = (uint)toExclusive - (uint)fromInclusive - 1;
             if (range == 0) return fromInclusive;
             uint mask = range;
@@ -162,7 +162,7 @@ namespace wan24.Crypto
             rng.FillBytes(data);
             if (data.IndexOf((byte)0) == -1) return;
             int i;
-            List<int> zeroIndex = new(),
+            List<int> zeroIndex = [],
                 newZeroIndex = null!;
             unchecked
             {
@@ -180,7 +180,7 @@ namespace wan24.Crypto
                     for (i = 0; i != zeroIndex.Count; i++)
                         if (buffer.Span[i] == 0)
                         {
-                            newZeroIndex ??= new();
+                            newZeroIndex ??= [];
                             newZeroIndex.Add(zeroIndex[i]);
                         }
                         else
@@ -203,7 +203,7 @@ namespace wan24.Crypto
             await rng.FillBytesAsync(data, cancellationToken).DynamicContext();
             if (data.IndexOf((byte)0) == -1) return;
             int i;
-            List<int> zeroIndex = new(),
+            List<int> zeroIndex = [],
                 newZeroIndex = null!;
             unchecked
             {
@@ -222,7 +222,7 @@ namespace wan24.Crypto
                     for (i = 0; i != zeroIndex.Count; i++)
                         if (buffer.Span[i] == 0)
                         {
-                            newZeroIndex ??= new();
+                            newZeroIndex ??= [];
                             newZeroIndex.Add(zeroIndex[i]);
                         }
                         else
