@@ -87,8 +87,7 @@ namespace wan24.Crypto
             _Timeout = stream.ReadLong();
             _Payload = stream.ReadULong();
             _MAC = new byte[MacHmacSha384Algorithm.MAC_LENGTH];
-            int red = stream.Read(_MAC);
-            if (red != _MAC.Length) throw new IOException($"Failed to read the MAC bytes (expected {STRUCT_LENGTH} bytes, red only {red} bytes)");
+            stream.ReadExactly(_MAC);
         }
 
         /// <summary>
@@ -141,7 +140,7 @@ namespace wan24.Crypto
         /// <summary>
         /// Time left until timeout
         /// </summary>
-        public TimeSpan Timeleft
+        public TimeSpan TimeLeft
         {
             get
             {
@@ -269,7 +268,7 @@ namespace wan24.Crypto
         /// <summary>
         /// Cast as serialized data
         /// </summary>
-        /// <param name="tt">Timeout tokenn</param>
+        /// <param name="tt">Timeout token</param>
         public static implicit operator byte[](in TimeoutToken tt) => tt.Serialize();
 
         /// <summary>
